pyscription/tests/examples/loadpipe/pyproject.toml

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "loadpipe"
version = "0.1.0"
description = "CLI pipeline for syncing Google Drive files with resume support and caching"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "AC DC PROCESSING" }]

# Minimal dependencies to boot the project; add new ones as needed.
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "PyYAML>=6.0.1",
    "fsspec>=2024.3.0"
]

[project.optional-dependencies]
gdrive = [
    "google-api-python-client>=2.147.0",
    "google-auth>=2.34.0",
    "google-auth-oauthlib>=1.2.0"
]
extras = [
    "requests>=2.32.0",
    "tqdm>=4.66.0",
    "tenacity>=8.4.0",
    "zstandard>=0.22.0"
]

[project.scripts]
lp = "loadpipe.cli:app"

[project.entry-points."fsspec.specs"]
gdrive = "loadpipe.filesystem:DriveFileSystem"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
